requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp_socks==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
frozenlist==1.4.1
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy==2.1.0
opencv-contrib-python==4.10.0.84
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-socks==2.5.1
requests==2.32.3
rich==13.8.0
setuptools==74.0.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
